rename user progress files and add route
