+++ server/server.js
@@ -28,7 +28,8 @@ app.use('/api/units', require('./routes/unitRoutes'))
 app.use('/api/users', require('./routes/userRoutes'))
 app.use('/api/videos', require('./routes/videoRoutes'))
 app.use('/api/xp', require('./routes/xpRoutes'));
-app.use('/api/lessonProgress', require('./routes/lessonProgressRoutes'))
+app.use('/api/lessonProgress', require('./routes/lessonProgressRoutes'));
+app.use('/api/userUnitProgress', require('./routes/userUnitProgressRoutes'))
 
 app.use(errorHandler)
